pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bot-moderator"
version = "0.1.0"
description = "Многофункциональный модератор Telegram, реализующий антиспам и сервисные возможности из инструкции DeFensy."
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "Codex"}]
dependencies = [
    "aiogram>=3.4.1",
    "pydantic>=2.5",
    "pydantic-settings>=2.2",
    "sqlmodel>=0.0.16",
    "aiosqlite>=0.19.0",
    "python-dateutil>=2.9.0",
    "pyyaml>=6.0.1",
    "orjson>=3.10.6"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "ruff>=0.6",
    "mypy>=1.11"
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "N", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
